A bucketful of pain
made uDBDisplay's dgbGrid dynamically resize depending on the largest text width found in each visible column by use of procedure FitGrid.

Moved delete after 1000 button to uDBDisplay.

Problems: FitGrid procedure does not save the width of each column, so when dbsDisplay is switched to another table, it is not called again and columns go back to old size problem.

Reason for not calling FitGrid upon each dbgDisplay change is because it might be taxing on the system, as it checks through the entire database.